OPENMEETINGS-266 almost fixed (audio should be implemented)
git-svn-id: https://svn.apache.org/repos/asf/incubator/openmeetings/trunk/singlewebapp@1362619 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/openmeetings/app/data/basic/Fieldmanagment.java
@@ -205,7 +205,7 @@ public class Fieldmanagment {
 			for (Object fl : query.getResultList()) {
 				Map<String, Object> map = new HashMap<String, Object>();
 				Object[] row = (Object[])fl;
-				map.put("id", (Long)row[0]);
+				map.put("id", row[0]);
 				map.put("value", performReplace((String)row[1], appName));
 				returnList.add(map);
 			}
